fixes bug: webapp requires a jangaroo source
[git-p4: depot-paths = "//coremedia/jangaroo/": change = 157652]

+++ maven-plugin/src/main/java/net/jangaroo/jooc/mvnplugin/WarPackageMojo.java
@@ -101,7 +101,7 @@ public class WarPackageMojo
       excludeFromWarPackaging();
 
       unpack(project, DefaultArtifact.SCOPE_RUNTIME, scriptDir);
-      if(compilerOutputRootDirectory != null) {
+      if(compilerOutputRootDirectory != null && compilerOutputRootDirectory.exists()) {
         FileUtils.copyDirectoryStructure(compilerOutputRootDirectory, scriptDir);
       }
     }
